requirements.txt
flask
pymongo
flask_pymongo
flask_cors
kafka-python